[mod] #8 클린아키텍처로 구조 변경
