+++ MEMO-Zi/app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools" >
+    xmlns:tools="http://schemas.android.com/tools">
 
     <uses-permission android:name="android.permission.INTERNET" />
 
@@ -43,7 +43,7 @@
         <activity
             android:name=".ui.memo.MemoActivity"
             android:exported="true"
-            android:screenOrientation="portrait" >
+            android:screenOrientation="portrait">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -51,11 +51,10 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".ui.diary.DiaryActivity"
+            android:name=".presentation.ui.diary.DiaryActivity"
             android:exported="false" />
         <activity
             android:name=".ui.setting.SettingActivity"
             android:exported="false" />
     </application>
-
 </manifest>
