+++ MEMO-Zi/app/src/main/java/com/memo_zi/presentation/model/DiaryItem.kt
@@ -1,0 +1,14 @@
+package com.memo_zi.presentation.model
+
+sealed class DiaryItem() {
+    data class Title(
+        val titleLogoImage: Int
+    ) : DiaryItem()
+
+    data class AllItem(
+        val diaryDate: String,
+        val diaryContent: String,
+        val diaryImgUrl: Int,
+        val diaryLocation: String,
+    ) : DiaryItem()
+}
