--- MEMO-Zi/app/src/main/java/com/memo_zi/presentation/ui/UiEx.kt
@@ -1,4 +1,0 @@
-package com.memo_zi.presentation.ui
-
-class UiEx {
-}
