+++ MEMO-Zi/app/src/main/java/com/memo_zi/presentation/ui/diary/DiaryCalendarFragment.kt
@@ -1,32 +1,26 @@
-package com.memo_zi.ui.diary
+package com.memo_zi.presentation.ui.diary
 
 import android.os.Bundle
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
 import android.widget.LinearLayout
 import androidx.fragment.app.viewModels
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.PagerSnapHelper
 import androidx.recyclerview.widget.RecyclerView
 import com.google.android.material.bottomsheet.BottomSheetBehavior
+import com.memo_zi.R
 import com.memo_zi.databinding.FragmentDiaryCalendarBinding
+import com.memo_zi.presentation.ui.diary.adapter.CalendarMonthAdapter
+import com.memo_zi.presentation.ui.diary.adapter.DiaryAdapter
 import com.memo_zi.util.binding.BindingFragment
 import timber.log.Timber
 
-class DiaryCalendarFragment : BindingFragment<FragmentDiaryCalendarBinding>() {
+class DiaryCalendarFragment :
+    BindingFragment<FragmentDiaryCalendarBinding>(R.layout.fragment_diary_calendar) {
     private val viewModel by viewModels<DiaryViewModel>()
     private lateinit var diaryAdapter: DiaryAdapter
-
     private lateinit var bottomSheetBehavior: BottomSheetBehavior<LinearLayout>
-
     private lateinit var recyclerView: RecyclerView
-
-    override fun getFragmentBinding(
-        inflater: LayoutInflater,
-        container: ViewGroup?
-    ): FragmentDiaryCalendarBinding =
-        FragmentDiaryCalendarBinding.inflate(inflater, container, false)
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
