+++ MEMO-Zi/app/src/main/java/com/memo_zi/presentation/ui/diary/DiaryFeedFragment.kt
@@ -1,21 +1,16 @@
-package com.memo_zi.ui.diary
+package com.memo_zi.presentation.ui.diary
 
 import android.os.Bundle
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
 import androidx.fragment.app.viewModels
+import com.memo_zi.R
 import com.memo_zi.databinding.FragmentDiaryFeedBinding
+import com.memo_zi.presentation.ui.diary.adapter.DiaryAdapter
 import com.memo_zi.util.binding.BindingFragment
 
-class DiaryFeedFragment : BindingFragment<FragmentDiaryFeedBinding>() {
+class DiaryFeedFragment : BindingFragment<FragmentDiaryFeedBinding>(R.layout.fragment_diary_feed) {
     private val viewModel by viewModels<DiaryViewModel>()
     private lateinit var diaryAdapter: DiaryAdapter
-
-    override fun getFragmentBinding(
-        inflater: LayoutInflater,
-        container: ViewGroup?
-    ): FragmentDiaryFeedBinding = FragmentDiaryFeedBinding.inflate(inflater, container, false)
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
